Reduce pain from relative imports.
The bundled version of Crypto really causes pain and suffering with strange
errors, type errors and bad signatures when mixed with other versions of
PyCrypto that might be on the search path.

I'm not sure this commit fixes everything that needs to be fixed, and there
is risk for breakage in parts not excersised by the unit tests. In particular,
pubkey generation as well as in the seemingly unused (looks broken) ElGamal
and DSA modules.